backend/requirements.txt
Flask
flask_pymongo
pymongo
flask-cors
PyJWT
bcrypt
